requirements.txt
aiofiles
python-dotenv
discord
aiohttp
validators